add authorization

+++ app/Http/Controllers/YapController.php
@@ -7,15 +7,18 @@ use Illuminate\Http\Request;
 
 class YapController extends Controller
 {
-    public function show(Yap $yap){
+    public function show(Yap $yap)
+    {
         return view('yaps.show', compact('yap'));
     }
 
-    public function store(){
-
+    public function store()
+    {
         $validated = request()->validate([
             'yap' => 'required|min:5|max:240'
         ]);
+
+        $validated['user_id'] = auth()->id();
 
         Yap::create($validated);
         // $yap = new Yap([
@@ -26,13 +29,20 @@ class YapController extends Controller
         return redirect()->route('dashboard')->with('success', 'Yapped Successfully!');
     }
 
-    public function edit(Yap $yap){
-
+    public function edit(Yap $yap)
+    {
+        if (auth()->id() !== $yap->user_id) {
+            abort(404, "You cannot do that");
+        }
         $editing = true;
         return view('yaps.show', compact('yap', 'editing'));
     }
 
-    public function update(Yap $yap){
+    public function update(Yap $yap)
+    {
+        if (auth()->id() !== $yap->user_id) {
+            abort(404, "You cannot do that");
+        }
         $validated = request()->validate([
             'yap' => 'required|min:5|max:240'
         ]);
@@ -44,9 +54,12 @@ class YapController extends Controller
         return redirect()->route('yaps.show', $yap->id)->with('success', 'Yap updated Successfully!');
     }
 
-    public function destroy(Yap $yap){
+    public function destroy(Yap $yap)
+    {
         // $yap = Yap::where('id', $id)->firstOrFail();
-
+        if (auth()->id() !== $yap->user_id) {
+            abort(404, "You cannot do that");
+        }
         $yap->delete();
 
         return redirect()->route('dashboard')->with('success', 'Unyapped Successfully!');
